v1.2.27 Refactor
- Reorganize folder structure
- Update test scripts
- Automatically set "track" based on "mode"
