+++ index.js
@@ -3,7 +3,7 @@
  * @ndaidong
 **/
 
-var main = require('./src/paypal-nvp-api');
+var main = require('./src/main');
 main.version = require('./package').version;
 
 module.exports = main;
