+++ package.json
@@ -1,5 +1,5 @@
 {
-  "version": "1.2.26",
+  "version": "1.2.27",
   "name": "paypal-nvp-api",
   "description": "Node.js wrapper for the Paypal Name-Value Pair — NVP ",
   "homepage": "https://www.npmjs.com/package/paypal-nvp-api",
