+++ src/helpers/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  formatCurrency: require('./formatCurrency'),
+  parse: require('./parse'),
+  stringify: require('./stringify')
+};
