+++ test/specs/config.js
@@ -1,7 +1,6 @@
 module.exports = {
-  rootDir: '../../../src/',
+  main: require('../../index'),
   mode: 'sandbox',
-  track: 'https://www.sandbox.paypal.com',
   username: 'alex-facilitator_api1.itravellocal.com',
   password: 'DYKNJZZE42ASU699',
   signature: 'A0aEilikhBmwfK.NlduDjCbsdgRdA8VDPMDksDhGsHmLQECu80Qtru09'
